Scripts/requirements.txt
scapy
smtplib
subprocess
python-dotenv
syslog
time